refactor: page controll component name
