+++ src/components/PageController/utils/getLottieOptions.ts
